+++ backend/internal/service/handler/auth/handler_test.go
@@ -103,7 +103,6 @@ func TestHandler_GuardianLogin(t *testing.T) {
 	for _, tt := range tests {
 		tt := tt // capture range variable for parallel
 		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
 			// Fixup for specific tests
 			if tt.name == "successful login" {
 				resp := tt.authResponse.(models.LoginResponse)
@@ -191,7 +190,6 @@ func TestHandler_ManagerLogin(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
 			if tt.name == "successful manager login" {
 				resp := tt.authResponse.(models.LoginResponse)
 				resp.User.ID = uuid.MustParse("b0eebc99-9c0b-4ef8-bb6d-6bb9bd380b22")
