+++ backend/internal/service/handler/organization/handler_test.go
@@ -742,6 +742,7 @@ func TestHandler_GetEventOccurrencesByOrganizationId(t *testing.T) {
 		name             string
 		id               string
 		mockSetup        func(*repomocks.MockOrganizationRepository)
+		mockS3Setup      func(*s3mocks.S3ClientMock)
 		wantErr          bool
 		statusCode       *int
 		messageSubstring *string
@@ -784,6 +785,10 @@ func TestHandler_GetEventOccurrencesByOrganizationId(t *testing.T) {
 					},
 				}, nil)
 			},
+			mockS3Setup: func(m *s3mocks.S3ClientMock) {
+				m.On("GeneratePresignedURL", mock.Anything, "events/robotics_workshop.jpg", mock.Anything).
+					Return("https://presigned.url/events/robotics_workshop.jpg", nil).Times(2)
+			},
 			wantErr: false,
 		},
 		{
@@ -797,7 +802,8 @@ func TestHandler_GetEventOccurrencesByOrganizationId(t *testing.T) {
 					"en-US",
 				).Return(make([]models.EventOccurrence, 0), nil)
 			},
-			wantErr: false,
+			mockS3Setup: func(m *s3mocks.S3ClientMock) {},
+			wantErr:     false,
 		},
 	}
 
@@ -811,6 +817,7 @@ func TestHandler_GetEventOccurrencesByOrganizationId(t *testing.T) {
 			tt.mockSetup(mockRepo)
 
 			mockS3 := createMockS3Client()
+			tt.mockS3Setup(mockS3)
 			handler := NewHandler(mockRepo, mockLocationRepo, mockS3)
 			ctx := context.Background()
