+++ frontend/apps/mobile/i18n/th.json
@@ -10,6 +10,7 @@
   "common": {
     "submit": "ส่ง",
     "cancel": "ยกเลิก",
+    "close": "ปิด",
     "loading": "กำลังโหลด...",
     "loadingMapData": "กำลังโหลดข้อมูลแผนที่...",
     "loadingEvents": "กำลังโหลดกิจกรรม...",
@@ -292,5 +293,23 @@
     "username": "ชื่อผู้ใช้",
     "changeImage": "เปลี่ยนรูปภาพ",
     "saveChanges": "บันทึกการเปลี่ยนแปลง"
+  },
+  "occurrence": {
+    "smiles": "/ 5 รอยยิ้ม",
+    "learnMore": "เรียนรู้เพิ่มเติม",
+    "reserve": "จอง",
+    "ages": "อายุ {{min}} ปี",
+    "agesRange": "อายุ {{min}} - {{max}} ปี",
+    "classTime": "คลาส {{time}}"
+  },
+  "reservation": {
+    "paymentFailed": "ชำระเงินไม่สำเร็จ กรุณาลองอีกครั้งหรือติดต่อฝ่ายสนับสนุน",
+    "completed": "เสร็จสิ้น",
+    "selectChildLabel": "เลือกเด็ก",
+    "noChildren": "ไม่พบข้อมูลเด็ก",
+    "termsNote": "เมื่อคลิก 'ชำระเงินตอนนี้' คุณอนุญาตให้ Skill Spark เรียกเก็บเงินจากบัตรของคุณในจำนวนเงินรวม บวกค่าธรรมเนียมที่เกี่ยวข้อง",
+    "terms": "ข้อกำหนดและเงื่อนไข",
+    "payAndReserve": "ชำระเงินและจอง",
+    "seeYouSoon": "พบกันเร็วๆ นี้!"
   }
 }
